Small fixes to unit prefabs in regards to unit type for class_system

+++ Game/Assets/Game/Scripts/Unit/class_system/ArcaneClass.cs
@@ -4,7 +4,7 @@ using System.Collections;
 [System.Serializable]
 public class Arcane: Entity
 {
-	public Arcane(int hp, int exhaust, int damage, float distance, float attack_range):
+	public Arcane(int hp, float exhaust, int damage, float distance, float attack_range):
 		base(hp, exhaust, damage, distance, attack_range, UnitType.Arcane)
 	{
 	}
