+++ Game/Assets/Game/Scripts/Unit/class_system/BraverClass.cs
@@ -4,7 +4,7 @@ using System.Collections;
 [System.Serializable]
 public class Braver: Entity
 {
-	public Braver(int hp, int exhaust, int damage, float distance, float attack_range):
+	public Braver(int hp, float exhaust, int damage, float distance, float attack_range):
 		base(hp, exhaust, damage, distance, attack_range, UnitType.Braver)
 	{
 	}
